home component broken out into home and search components, helper for api calls created
